Consolidate string/text conversions outside of API
