+++ osquery/core/conversions.cpp
@@ -61,9 +61,9 @@ std::string base64Encode(const std::string& unencoded) {
     return std::string();
   }
 
-  size_t writePaddChars = (3U-unencoded.length()%3U)%3U;
+  size_t writePaddChars = (3U - unencoded.length() % 3U) % 3U;
   std::string base64(it_base64(unencoded.begin()), it_base64(unencoded.end()));
-  base64.append(writePaddChars,'=');
+  base64.append(writePaddChars, '=');
   os << base64;
   return os.str();
 }
@@ -75,5 +75,45 @@ bool isPrintable(const std::string& check) {
     }
   }
   return true;
+}
+
+std::vector<std::string> split(const std::string& s, const std::string& delim) {
+  std::vector<std::string> elems;
+  boost::split(elems, s, boost::is_any_of(delim));
+  auto start =
+      std::remove_if(elems.begin(), elems.end(), [](const std::string& s) {
+        return s.size() == 0;
+      });
+  elems.erase(start, elems.end());
+  for (auto& each : elems) {
+    boost::algorithm::trim(each);
+  }
+  return elems;
+}
+
+std::vector<std::string> split(const std::string& s,
+                               const std::string& delim,
+                               size_t occurences) {
+  // Split the string normally with the required delimiter.
+  auto content = split(s, delim);
+  // While the result split exceeds the number of requested occurrences, join.
+  std::vector<std::string> accumulator;
+  std::vector<std::string> elems;
+  for (size_t i = 0; i < content.size(); i++) {
+    if (i < occurences) {
+      elems.push_back(content.at(i));
+    } else {
+      accumulator.push_back(content.at(i));
+    }
+  }
+  // Join the optional accumulator.
+  if (accumulator.size() > 0) {
+    elems.push_back(join(accumulator, delim));
+  }
+  return elems;
+}
+
+std::string join(const std::vector<std::string>& s, const std::string& tok) {
+  return boost::algorithm::join(s, tok);
 }
 }
