+++ osquery/core/conversions.h
@@ -14,6 +14,7 @@
 
 #include <memory>
 #include <string>
+#include <vector>
 
 #include <boost/bind.hpp>
 #include <boost/shared_ptr.hpp>
@@ -49,6 +50,63 @@ typename boost::shared_ptr<T> std_to_boost_shared_ptr(
     typename std::shared_ptr<T> const& p) {
   return boost::shared_ptr<T>(p.get(), boost::bind(&do_release_std<T>, p, _1));
 }
+
+/**
+ * @brief Split a given string based on an optional delimiter.
+ *
+ * If no delimiter is supplied, the string will be split based on whitespace.
+ *
+ * @param s the string that you'd like to split
+ * @param delim the delimiter which you'd like to split the string by
+ *
+ * @return a vector of strings split by delim.
+ */
+std::vector<std::string> split(const std::string& s,
+                               const std::string& delim = "\t ");
+
+/**
+ * @brief Split a given string based on an delimiter.
+ *
+ * @param s the string that you'd like to split.
+ * @param delim the delimiter which you'd like to split the string by.
+ * @param occurrences the number of times to split by delim.
+ *
+ * @return a vector of strings split by delim for occurrences.
+ */
+std::vector<std::string> split(const std::string& s,
+                               const std::string& delim,
+                               size_t occurences);
+
+/**
+ * @brief In-line replace all instances of from with to.
+ *
+ * @param str The input/output mutable string.
+ * @param from Search string
+ * @param to Replace string
+ */
+inline void replaceAll(std::string& str,
+                       const std::string& from,
+                       const std::string& to) {
+  if (from.empty()) {
+    return;
+  }
+
+  size_t start_pos = 0;
+  while ((start_pos = str.find(from, start_pos)) != std::string::npos) {
+    str.replace(start_pos, from.length(), to);
+    start_pos += to.length();
+  }
+}
+
+/**
+ * @brief Join a vector of strings using a tokenizer.
+ *
+ * @param s the string that you'd like to split.
+ * @param tok a token glue.
+ *
+ * @return a joined string.
+ */
+std::string join(const std::vector<std::string>& s, const std::string& tok);
 
 /**
  * @brief Decode a base64 encoded string.
@@ -120,6 +178,43 @@ inline std::string unescapeUnicode(const std::string& escaped) {
   return unescaped;
 }
 
+/**
+ * @brief In-line helper function for use with utf8StringSize
+ */
+template <typename _Iterator1, typename _Iterator2>
+inline size_t incUtf8StringIterator(_Iterator1& it, const _Iterator2& last) {
+  if (it == last) {
+    return 0;
+  }
+
+  size_t res = 1;
+  for (++it; last != it; ++it, ++res) {
+    unsigned char c = *it;
+    if (!(c & 0x80) || ((c & 0xC0) == 0xC0)) {
+      break;
+    }
+  }
+
+  return res;
+}
+
+/**
+ * @brief Get the length of a UTF-8 string
+ *
+ * @param str The UTF-8 string
+ *
+ * @return the length of the string
+ */
+inline size_t utf8StringSize(const std::string& str) {
+  size_t res = 0;
+  std::string::const_iterator it = str.begin();
+  for (; it != str.end(); incUtf8StringIterator(it, str.end())) {
+    res++;
+  }
+
+  return res;
+}
+
 #ifdef DARWIN
 /**
  * @brief Convert a CFStringRef to a std::string.
@@ -139,5 +234,4 @@ std::string stringFromCFAbsoluteTime(const CFDataRef& cf_abstime);
 
 std::string stringFromCFData(const CFDataRef& cf_data);
 #endif
-
 }
