+++ osquery/tables/system/crontab.cpp
@@ -17,6 +17,8 @@
 #include <osquery/filesystem.h>
 #include <osquery/logger.h>
 
+#include "osquery/core/conversions.h"
+
 namespace osquery {
 namespace tables {
