+++ osquery/tables/system/darwin/homebrew_packages.cpp
@@ -17,6 +17,8 @@
 #include <osquery/logger.h>
 #include <osquery/tables.h>
 
+#include "osquery/core/conversions.h"
+
 namespace fs = boost::filesystem;
 namespace pt = boost::property_tree;
