+++ osquery/tables/system/darwin/launchd.cpp
@@ -18,6 +18,8 @@
 #include <osquery/logger.h>
 #include <osquery/tables.h>
 
+#include "osquery/core/conversions.h"
+
 namespace fs = boost::filesystem;
 namespace pt = boost::property_tree;
