+++ osquery/tables/system/darwin/os_version.cpp
@@ -13,6 +13,8 @@
 #include <osquery/sql.h>
 #include <osquery/tables.h>
 
+#include "osquery/core/conversions.h"
+
 namespace osquery {
 namespace tables {
