+++ osquery/tables/system/linux/kernel_modules.cpp
@@ -10,7 +10,6 @@
 
 #include <fstream>
 
-#include <boost/algorithm/string/split.hpp>
 #include <boost/algorithm/string/trim.hpp>
 
 #include <osquery/core.h>
@@ -18,10 +17,12 @@
 #include <osquery/logger.h>
 #include <osquery/tables.h>
 
+#include "osquery/core/conversions.h"
+
 namespace osquery {
 namespace tables {
 
-const std::string kKernelModulePath = "/proc/modules";
+static const std::string kKernelModulePath = "/proc/modules";
 
 QueryData genKernelModules(QueryContext& context) {
   QueryData results;
@@ -41,9 +42,9 @@ QueryData genKernelModules(QueryContext& context) {
   auto module_info = std::string(std::istreambuf_iterator<char>(fd),
                                  std::istreambuf_iterator<char>());
 
-  for (const auto& module : split(module_info, "\n")) {
+  for (const auto& module : osquery::split(module_info, "\n")) {
     Row r;
-    auto module_info = split(module, " ");
+    auto module_info = osquery::split(module, " ");
     if (module_info.size() < 6) {
       // Interesting error case, this module line is not well formed.
       continue;
