+++ osquery/tables/system/linux/system_info.cpp
@@ -14,6 +14,7 @@
 #include <osquery/tables.h>
 #include <osquery/sql.h>
 
+#include "osquery/core/conversions.h"
 #include "osquery/tables/system/linux/smbios_utils.h"
 
 namespace osquery {
